Move matchmaking utilities file to matchmaking feature folder
