+++ frontend/src/features/matchmaking/utils.ts
@@ -1,0 +1,12 @@
+import { getUsername } from "../../utils/apiUtils";
+
+async function getAssignedColor(whitePlayer: string, blackPlayer: string) {
+	const username = await getUsername();
+
+	const isPlayingWhite = username === whitePlayer;
+	const assignedColor = isPlayingWhite ? "white" : "black";
+
+	return assignedColor;
+}
+
+export { getAssignedColor };
